pyproject.toml
[project]
name = "controlnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.13.0",
    "controlnet-aux>=0.0.10",
    "diffusers==0.27.2",
    "gradio>=6.10.0",
    "huggingface-hub>=1.7.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "torch-directml>=0.2.5.dev240914",
    "torchvision>=0.19.1",
    "transformers>=5.3.0",
]
